Reduced PC signature to 8-bits and hash table to 256 entries

+++ inc/block.h
@@ -34,7 +34,7 @@ struct cache_block {
     // ➕ ML Features
   uint8_t hits_since_insertion = 0;  // 2–3 bits, store hit count since inserted
   bool reused = false;               // Set if accessed at least once after fill
-  uint16_t pc_signature = 0;         // Hashed PC, 10–14 bits
+  uint8_t pc_signature = 0;         // Hashed PC, 10–14 bits
 
   // Optional: track recency (if not done by RRIP or LRU policy already)
   uint8_t recency = 0;
